Refactored code structure, endpoints handing label changes in document have been moved to a different file.
